smsjwplatform removed
All contents of smsjwplatform moved to mediaplatform_jwp. smsjwplatform left with migrations in order to migrate the data from one app to the other but it should be scheduled to be removed once this has finished and we have migrated all databases.
