+++ api/serializers.py
@@ -9,10 +9,8 @@ from django.urls import reverse
 from django.utils.http import urlencode
 from rest_framework import serializers
 
-from smsjwplatform import jwplatform
 from mediaplatform import models as mpmodels
-from mediaplatform_jwp import management
-
+from mediaplatform_jwp import managementapi as management, jwplatform
 
 LOG = logging.getLogger(__name__)
