+++ doc/mediaplatform_jwp.rst
@@ -1,0 +1,49 @@
+JWPlatform Integration Application
+==================================
+
+.. automodule:: mediaplatform_jwp
+
+Settings
+--------
+
+.. automodule:: mediaplatform_jwp.defaultsettings
+    :members:
+
+Models
+------
+
+.. automodule:: mediaplatform_jwp.models
+    :members:
+
+Management commands
+-------------------
+
+jwpfetch
+````````
+
+.. automodule:: mediaplatform_jwp.management.commands.jwpfetch
+
+JWPlatform API
+--------------
+
+.. automodule:: mediaplatform_jwp.jwplatform
+    :members:
+    :member-order: bysource
+
+ACLs
+----
+
+.. automodule:: mediaplatform_jwp.acl
+    :members:
+
+Application configuration
+-------------------------
+
+.. automodule:: mediaplatform_jwp.apps
+    :members:
+
+System checks
+-------------
+
+.. automodule:: mediaplatform_jwp.systemchecks
+    :members:
