--- doc/smsjwplatform.rst
@@ -1,70 +1,0 @@
-JWPlatform Integration Application
-==================================
-
-.. automodule:: smsjwplatform
-
-Settings
---------
-
-.. automodule:: smsjwplatform.defaultsettings
-    :members:
-
-.. automodule:: mediaplatform_jwp.defaultsettings
-    :members:
-
-Views
------
-
-.. automodule:: smsjwplatform.views
-    :members:
-
-Models
-------
-
-.. automodule:: smsjwplatform.models
-    :members:
-
-.. automodule:: mediaplatform_jwp.models
-    :members:
-    :member-order: bysource
-
-
-Management commands
--------------------
-
-jwpfetch
-````````
-
-.. automodule:: smsjwplatform.management.commands.jwpfetch
-
-
-URLs
-----
-
-.. automodule:: smsjwplatform.urls
-    :members:
-
-JWPlatform API
---------------
-
-.. automodule:: smsjwplatform.jwplatform
-    :members:
-    :member-order: bysource
-
-ACLs
-----
-
-.. automodule:: smsjwplatform.acl
-    :members:
-
-Application configuration
--------------------------
-
-.. automodule:: smsjwplatform.apps
-    :members:
-
-System checks
--------------
-
-.. automodule:: smsjwplatform.systemchecks
-    :members:
