+++ mediaplatform_jwp/managementapi.py
@@ -2,7 +2,7 @@
 Interactions with the JWP management API.
 
 """
-from smsjwplatform import jwplatform as jwp
+from mediaplatform_jwp import jwplatform as jwp
 
 from . import models
 from . import upload
