+++ smsjwplatform/__init__.py
@@ -2,4 +2,3 @@
 The :py:mod:`smsjwplatform` application provides integration with the JWPlatform API.
 
 """
-default_app_config = 'smsjwplatform.apps.SMSJWPlatformConfig'
