--- smsjwplatform/tests/test_views.py
@@ -1,6 +1,0 @@
-"""
-Tests for views.
-
-"""
-
-# No views at present
