--- smsjwplatform/views.py
@@ -1,5 +1,0 @@
-"""
-Views
-"""
-
-# No views at present
